% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/platereading.R
\docType{data}
\name{buchworm}
\alias{buchworm}
\title{The Buchanan model, paramterised for worms}
\format{\preformatted{Class 'formula' length 3 od ~ od0 + (time >= lag) * (time <= (lag + (odmax - od0) * log(10)/mumax)) * mumax * (time - lag)/log(10) + (time >=      lag) * (time > (lag + (odmax - od0) * log(10)/mumax)) * (odmax - od0)
  ..- attr(*, ".Environment")=<environment: namespace:platereading> 
}}
\usage{
buchworm
}
\description{
The model as described by Buchanan et al (1997) -
When is simple good enough: a comparison of the Gompertz, Baranyi, and three-phase linear models for fitting bacterial growth curves.
}
\keyword{datasets}

