% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/platereading.R
\name{fitbuch}
\alias{fitbuch}
\title{Fit a single directory of data (output by readonedir) with the buch function}
\usage{
fitbuch(x, y = 1)
}
\arguments{
\item{x}{A Dataframe produced by readonedir}

\item{y}{an indication of whether to omit t=0 or not, defaults to 1, remove}
}
\value{
A dataframe containing the fitted data
}
\description{
Fit a single directory of data (output by readonedir) with the buch function
}

