% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereading.R
\name{looper}
\alias{looper}
\title{Read in a directory containing multiple plates of data}
\usage{
looper(x, lag1 = 35, mumax1 = 0.025, od01 = 0.25, odmax1 = 0.95)
}
\arguments{
\item{x}{A Directory containing all data and required outputfiles}
}
\value{
A dataframe containing the data fit and sorted by well
}
\description{
Read in a directory containing multiple plates of data
}

