% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/platereading.R
\name{namer}
\alias{namer}
\title{Name output from a directory containing multiple plates of data}
\usage{
namer(x, y)
}
\arguments{
\item{x}{A dataframe containing the output of the looper function}

\item{y}{A file containing all names of wells in plates}
}
\value{
A dataframe containing the data fit and sorted by well
}
\description{
Name output from a directory containing multiple plates of data
}

