% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/platereading.R
\name{plateshiny}
\alias{plateshiny}
\title{A shiny app for interactive production of plots and outputs}
\usage{
plateshiny(directory)
}
\arguments{
\item{directory}{the directory containing data and strain names}
}
\value{
a shiny app
}
\description{
A shiny app for interactive production of plots and outputs
}

