% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/platereading.R
\name{plotter}
\alias{plotter}
\title{Plot a single well with its fit}
\usage{
plotter(x, well, lag1 = 35, mumax1 = 0.025, od01 = 0.25, odmax1 = 0.95)
}
\arguments{
\item{x}{dataframe containing runs}

\item{well}{the well to choose}
}
\value{
plots of the well
}
\description{
Plot a single well with its fit
}

