% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/platereading.R
\name{plyrfit}
\alias{plyrfit}
\title{Fit multiple directories of data (output by readonedir) with the buch function}
\usage{
plyrfit(x, lag1 = 35, mumax1 = 0.025, od01 = 0.25, odmax1 = 0.95)
}
\arguments{
\item{x}{A Dataframe produced by outputs of readonedir}
}
\value{
A dataframe containing the fitted data
}
\description{
Fit multiple directories of data (output by readonedir) with the buch function
}

