% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/platereading.R
\name{readonedir}
\alias{readonedir}
\title{Read in a single plate directory of data}
\usage{
readonedir(x)
}
\arguments{
\item{x}{A Directory containing one plate run}
}
\value{
A dataframe containing the data sorted by well and time
}
\description{
Read in a single plate directory of data
}

