% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/platereading.R
\docType{data}
\name{buchworm}
\alias{buchworm}
\title{The Buchanan model for three phase growth}
\format{an equation, with od modelled as a response of od0, lag, mumax and odmax}
\usage{
buchworm
}
\description{
The model as described by	Buchanan, R. L., Whiting, R. C. & Damert,
W. C. 1997 When is simple good enough: A comparison of the Gompertz, Baranyi,
and three-phase linear models for fitting bacterial growth curves.
Food Microbiol. 14, 313–326.
}
\keyword{datasets}

