% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/platereading.R
\name{looper}
\alias{looper}
\title{Read in a directory containing multiple plates of data}
\usage{
looper(x)
}
\arguments{
\item{x}{A Directory containing all data and required outputfiles}
}
\value{
A dataframe containing the data fit and sorted by well
}
\description{
Read in a directory containing multiple plates of data
}

