% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platereading.R
\name{plotlots}
\alias{plotlots}
\title{Plot all representatives of a strain with their fits and residuals}
\usage{
plotlots(x, y, z, lag1 = 35, mumax1 = 0.025, od01 = 0.25, odmax1 = 0.95)
}
\arguments{
\item{x}{A directory containing multiple plate runs}

\item{y}{A strain to plot out}

\item{z}{The file containing strain names}
}
\value{
plots of each well of the required strain
}
\description{
Plot all representatives of a strain with their fits and residuals
}

