% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/platereading.R
\name{plotter}
\alias{plotter}
\title{Plot a single well with its fit}
\usage{
plotter(x, well)
}
\arguments{
\item{x}{dataframe containing runs}

\item{well}{the well to choose}
}
\value{
plots of the well
}
\description{
Plot a single well with its fit
}

