% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/platereading.R
\name{plyrfit}
\alias{plyrfit}
\title{Fit multiple directories of data (output by readonedir) with the buch function}
\usage{
plyrfit(x)
}
\arguments{
\item{x}{A Dataframe produced by outputs of readonedir}
}
\value{
A dataframe containing the fitted data
}
\description{
Fit multiple directories of data (output by readonedir) with the buch function
}

